+++ src/components/blogs/Blog.css
@@ -13,6 +13,7 @@
 /* Create two unequal columns that floats next to each other */
 /* Left column */
 .leftcolumn {
+  padding: 2%;
   margin: 0 auto;
   width: 60%;
 }
